requirements.txt
requests
bs4
pandas
datetime 
dateutil.relativedelta 
re
pyyaml
